Seperate validator functions to a module.
Refactor to have validator retrieval and online check functions in a
seperate module. The result will return the list of an object with
validator infos and their last up status. These will be called from the main function
and passed to the emailer module.

Also npm install nodemailer and dotenv