Test send email.
Currently gets the validators that are checked for uptime and sends an
email. The email is plaintext formatted but dynamically gets contact
info. The contact info is not dnamic but uses a dummy contact. This will
be replaced by a mapping from the contacts exported from Cobra at first.
Later throgh automatic access to Cobra DB.